+++ tests/testthat/test-check_wget.R
@@ -1,0 +1,8 @@
+library(testthat)
+
+test_that("test if wget is installed", {
+  wget= check_wget()
+
+  # check the result
+  expect_true(is.logical(wget))
+})