+++ NEWS.md
@@ -1,6 +1,11 @@
+# lulab.utils 0.0.4
+
++ `check_wget()` & `use_wget()`
+
 # lulab.utils 0.0.3 (2024-09-30)
 
 + `test_mirror()`
+
 # lulab.utils 0.0.2 (2024-08-30)
 
 + `check_cha()`